requirements.txt
paho-mqtt==1.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-environ==0.4.5
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
httpx==0.11.1
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
whitenoise==5.0.1
black==19.10b0
boto3==1.11.9
